working mostly generic controllers
