+++ app/controllers/StoragesController.scala
@@ -5,7 +5,7 @@ import models.{StorageEntry, StorageEntryRow, StorageSerializer}
 import play.api.Configuration
 import play.api.db.slick.DatabaseConfigProvider
 import play.api.libs.json._
-import play.api.mvc.{Action, BodyParsers}
+import play.api.mvc.{Action, BodyParsers, Request}
 import slick.driver.JdbcProfile
 import slick.lifted.TableQuery
 import slick.driver.PostgresDriver.api._
@@ -16,44 +16,19 @@ import scala.concurrent.ExecutionContext.Implicits.global
 
 class StoragesController @Inject()
     (configuration: Configuration, dbConfigProvider: DatabaseConfigProvider)
-    extends GenericDatabaseObjectController with StorageSerializer {
+    extends GenericDatabaseObjectController[StorageEntry] with StorageSerializer {
+
   val dbConfig = dbConfigProvider.get[JdbcProfile]
 
+  override def selectall = dbConfig.db.run(
+    TableQuery[StorageEntryRow].result.asTry //simple select *
+  )
 
-  override def list = Action.async {
-    dbConfig.db.run(
-      TableQuery[StorageEntryRow].result.asTry //simple select *
-    ).map({
-      case Success(result)=>Ok(Json.obj("status"->"ok","result"->result))
-      case Failure(error)=>InternalServerError(Json.obj("status"->"error","detail"->error.toString))
-    })
-  }
+  override def jstranslate(result: Seq[StorageEntry]) = result.asInstanceOf[Seq[StorageEntry]]  //implicit translation should handle this
 
-  override def create = Action.async(BodyParsers.parse.json) { request =>
-    request.body.validate[StorageEntry].fold(
-      errors => {
-        Future(BadRequest(Json.obj("status"->"error","detail"->JsError.toJson(errors))))
-      },
-      storageEntry => {
-        dbConfig.db.run(
-          (TableQuery[StorageEntryRow] returning TableQuery[StorageEntryRow].map(_.id) += storageEntry).asTry
-        ).map({
-          case Success(result)=>Ok(Json.obj("status" -> "ok", "detail" -> "added", "id" -> result))
-          case Failure(error)=>InternalServerError(Json.obj("status"->"error", "detail"->error.toString))
-        }
-        )
-      }
-    )
-  }
+  override def insert(storageEntry: StorageEntry) = dbConfig.db.run(
+    (TableQuery[StorageEntryRow] returning TableQuery[StorageEntryRow].map(_.id) += storageEntry).asTry
+  )
 
-  override def update(id: Int) = Action.async(BodyParsers.parse.json) { request =>
-    request.body.validate[StorageEntry].fold(
-      errors=>Future(BadRequest(Json.obj("status"->"error","detail"->JsError.toJson(errors)))),
-      StorageEntry=>Future(Ok(""))
-    )
-  }
-
-  override def delete(id: Int) = Action.async(BodyParsers.parse.json) { request =>
-    Future(Ok(""))
-  }
+  override def validate(request:Request[JsValue]) = request.body.validate[StorageEntry]
 }
